#107 - Split active-user search into its own controller and service
Apply single-responsibility to the management controllers. Extract the active-user search out of group management into a dedicated IUserSearchService/UserSearchService<TUser> and an abstract BaseUserSearchController (route 'UserSearch'). BaseGroupManagementController now only manages groups.

Also remove the unused GroupManagement/{id}/UserList endpoint (and its Wolverine counterpart and the GetGroupUsersAsync service method): group members are already returned by GetGroupAsync, and no cauca-management code consumes the separate listing. Tests, DI registration and README updated; 269 backend tests pass.
