+++ Cause.SecurityManagement.Integration.Tests/Management/GroupManagementApiServiceTests.cs
@@ -113,72 +113,6 @@ public class GroupManagementApiServiceTests : IntegrationTestBase
         (await Service.GetGroupAsync(Guid.NewGuid())).Should().BeNull();
     }
 
-    [Test]
-    public async Task WhenGroupHasMembers_GetGroupUsers_ShouldReturnThem()
-    {
-        var user = SeedUser("Margaret", "Hamilton");
-        var groupId = Guid.NewGuid();
-        await Service.SaveGroupAsync(new GroupDto
-        {
-            Id = groupId,
-            Name = $"Group_{Guid.NewGuid():N}",
-            Users = [new GroupUserDto { Id = user.Id }],
-        });
-
-        var members = await Service.GetGroupUsersAsync(groupId);
-
-        members.Should().ContainSingle();
-        members[0].FirstName.Should().Be("Margaret");
-        members[0].LastName.Should().Be("Hamilton");
-    }
-
-    [Test]
-    public async Task WhenSearching_SearchUsers_ShouldMatchFirstOrLastNameCaseInsensitiveAndExcludeIds()
-    {
-        var token = $"Z{Guid.NewGuid():N}";
-        var first = SeedUser($"{token}first", "Smith");
-        var second = SeedUser("John", $"{token}last");
-        SeedUser("Unrelated", "Person");
-
-        var result = await Service.SearchUsersAsync(new UserSearchRequestDto
-        {
-            Query = token.ToUpper(),
-            Skip = 0,
-            Top = 50,
-            ExcludedUserIds = [second.Id],
-        });
-
-        result.TotalCount.Should().Be(1);
-        result.Items.Should().ContainSingle().Which.Id.Should().Be(first.Id);
-    }
-
-    [Test]
-    public async Task WhenSearching_SearchUsers_ShouldPageAndReportTotalBeforePaging()
-    {
-        var token = $"Z{Guid.NewGuid():N}";
-        SeedUser($"{token}a", "User");
-        SeedUser($"{token}b", "User");
-        SeedUser($"{token}c", "User");
-
-        var result = await Service.SearchUsersAsync(new UserSearchRequestDto { Query = token, Skip = 1, Top = 1 });
-
-        result.TotalCount.Should().Be(3);
-        result.Items.Should().ContainSingle();
-    }
-
-    [Test]
-    public async Task WhenSearching_SearchUsers_ShouldOnlyReturnActiveUsers()
-    {
-        var token = $"Z{Guid.NewGuid():N}";
-        var active = SeedUser($"{token}active", "User");
-        SeedUser($"{token}inactive", "User", isActive: false);
-
-        var result = await Service.SearchUsersAsync(new UserSearchRequestDto { Query = token, Skip = 0, Top = 50 });
-
-        result.TotalCount.Should().Be(1);
-        result.Items.Should().ContainSingle().Which.Id.Should().Be(active.Id);
-    }
-
     private TestUser SeedUser(string firstName, string lastName, bool isActive = true)
     {
         var user = new TestUser
